feat: change hooks to another file
